Allow reporting replication as unhealthy initially
Replication health reporting is based on Eventuate's
Available/Unavailable messages. However those are not sent when a
connection to a remote endpoint cannot be established at all (the ids
of these log respectively the corresponding endpoint is not known
until after a successful connect).

Now the ReplicationHealthMonitor can take an optional set of log-ids
for logs whose replication status shall be reported as unhealthy,
if the initial connection cannot be established.

Closes #25